#5 Update Health to Be Int
Realized what's the point of having half a hit point? I was thinking like "oh when we do division and multiplication and whatnot we'll get an EXACT attack value etc." but like why. Simply round to the nearest value.

+++ Scenes/Shared_Components/HealthComponent.cs
@@ -7,9 +7,9 @@ public partial class HealthComponent : Node
     public delegate void HealthDepletedEventHandler();
 
     [Export]
-    public double MaxHealth = 100.0;
+    public int MaxHealth = 100;
 
-    public double CurrentHealth;
+    public int CurrentHealth;
 
     public override void _EnterTree()
     {
@@ -18,11 +18,11 @@ public partial class HealthComponent : Node
     }
 
 
-    public void TakeDamage(double damageAmount)
+    public void TakeDamage(int damageAmount)
     {
         CurrentHealth -= damageAmount;
-        
-        if(CurrentHealth <= 0.0)
+
+        if(CurrentHealth <= 0)
         {
             EmitSignal(SignalName.HealthDepleted);
         }
